# Remove array copy when decoding UTF-8 field.