--- a/zipkin/src/main/java/zipkin2/internal/Buffer.java
+++ b/zipkin/src/main/java/zipkin2/internal/Buffer.java
@@ -337,7 +337,7 @@ public abstract class Buffer {
 
   public abstract int pos();
 
-  public abstract byte[] toByteArray();
+  public abstract byte[] toByteArrayUnsafe();
 
   public interface Writer<T> {
     int sizeInBytes(T value);