--- a/zipkin/src/main/java/zipkin2/internal/Dependencies.java
+++ b/zipkin/src/main/java/zipkin2/internal/Dependencies.java
@@ -78,7 +78,7 @@ public final class Dependencies {
   public ByteBuffer toThrift() {
     Buffer buffer = Buffer.allocate(sizeInBytes());
     write(buffer);
-    return ByteBuffer.wrap(buffer.toByteArray());
+    return ByteBuffer.wrap(buffer.toByteArrayUnsafe());
   }
 
   int sizeInBytes() {