In the benchmark, this affects 7 fields instead of 3, so actually has a larger impact than #2582 (this PR does not include that one).

Also renamed `toByteArray` to include `Unsafe` since it is.

After
```
Benchmark                                                                    Mode    Cnt        Score        Error   Units
ProtoCodecBenchmarks.bytes_zipkinDecoder                                   sample  14048     1068.003 ±      5.292   us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p0.00         sample             919.552                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p0.50         sample            1010.688                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p0.90         sample            1212.416                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p0.95         sample            1347.584                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p0.99         sample            2002.944                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p0.999        sample            2506.351                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p0.9999       sample            4236.627                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p1.00         sample            4530.176                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.alloc.rate                    sample     15     1097.823 ±     29.917  MB/sec
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.alloc.rate.norm               sample     15  1847161.027 ±     18.455    B/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.churn.G1_Eden_Space           sample     15     1110.072 ±     85.859  MB/sec
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.churn.G1_Eden_Space.norm      sample     15  1866332.749 ± 107577.650    B/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.churn.G1_Old_Gen              sample     15        0.803 ±      0.645  MB/sec
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.churn.G1_Old_Gen.norm         sample     15     1350.408 ±   1090.177    B/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.churn.G1_Survivor_Space       sample     15        0.488 ±      0.628  MB/sec
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.churn.G1_Survivor_Space.norm  sample     15      808.818 ±   1034.386    B/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.count                         sample     15      116.000               counts
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.time                          sample     15      103.000                   ms
```

Before
```
Benchmark                                                                    Mode    Cnt        Score        Error   Units
ProtoCodecBenchmarks.bytes_zipkinDecoder                                   sample  13504     1110.811 ±      5.541   us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p0.00         sample             942.080                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p0.50         sample            1060.864                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p0.90         sample            1269.760                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p0.95         sample            1376.256                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p0.99         sample            2080.154                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p0.999        sample            2699.264                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p0.9999       sample            5006.438                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:bytes_zipkinDecoder·p1.00         sample            5652.480                us/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.alloc.rate                    sample     15     1192.757 ±     75.891  MB/sec
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.alloc.rate.norm               sample     15  2087164.817 ±     22.574    B/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.churn.G1_Eden_Space           sample     15     1189.685 ±     95.704  MB/sec
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.churn.G1_Eden_Space.norm      sample     15  2081750.627 ± 103259.490    B/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.churn.G1_Old_Gen              sample     15        1.142 ±      0.611  MB/sec
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.churn.G1_Old_Gen.norm         sample     15     2024.829 ±   1101.138    B/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.churn.G1_Survivor_Space       sample     15        1.553 ±      0.581  MB/sec
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.churn.G1_Survivor_Space.norm  sample     15     2735.052 ±   1038.158    B/op
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.count                         sample     15      116.000               counts
ProtoCodecBenchmarks.bytes_zipkinDecoder:·gc.time                          sample     15      110.000                   ms
```